requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-telegram-bot==13.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
tornado==6.1
transmission-rpc==3.2.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
